move ownership related tests
